Migrating Fake test objects to the new approach (sharing objects between unit and UI Tests) (DATA layer)
